fix: rename after clearing ts cache and restarting to handle casing.
